pyproject.toml
[tool.poetry]
name = "lazarus"
version = "2.0.0"
description = "High Availability Reddit Memes Analyzer, for fun (and passing Distribuidos I)."
authors = ["CrossNox <imermet@fi.uba.ar>", "FdelMazo <fdelmazo@fi.uba.ar>", "JulianVentura <jventura@fi.uba.ar>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
typer = "^0.4.1"
youconfigme = "^0.6.15"
mypy-extensions = "^0.4.3"
requests = "^2.27.1"
types-requests = "^2.27.27"
pika = "^1.2.1"
docker = "^5.0.3"
pyzmq = "^23.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
isort = "^5.10.1"
mypy = "^0.950"
pre-commit = "^2.19.0"
flake8 = "^4.0.1"
kaggle = "^1.5.12"
pandas = "^1.4.2"
pycln = "^1.3.3"
pylint = "^2.14.4"
black = "^22.6.0"

[tool.poetry.scripts]
lazarus = "lazarus.cli.cli:app"
virgilius = "lazarus.client.cli:app"
mom_test = "tests.mom.app:main"
protocol_test = "tests.protocol.app:main"

[tool.black]
line-length = 88

[tool.mypy]
ignore_missing_imports = true
follow_imports = "normal"

[tool.isort]
line_length=88
indent='    '
multi_line_output=3
length_sort=true
use_parentheses=true
include_trailing_comma=true
force_sort_within_sections=true
known_first_party="lazarus"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
